Add star rating code reference
-I was not able to find a code to display the stars properly on my own,
I had to take this code, because most codes I have tried did not want to
display the stars due to some JQuery library problems, it only displayed
radio buttons.
-I also added a function to reset the rating stars, to allow users to
change the value of the stars, or else they could only choose once and
refresh the page and lose all the comments written priorly.
-There was a malfunctionning with the textarea, text cursor now starts
on left.